# PyArray conversion speedups and PyArrayFromBuffer