--- a/src/numpy.jl
+++ b/src/numpy.jl
@@ -102,7 +102,7 @@ end
 # the values of these seem to have been stable for some time, and
 # the NumPy developers seem to have some awareness of binary compatibility
 
-# NPY_TYPES:
+# NumPy Types:
 
 const NPY_BOOL = Int32(0)
 const NPY_BYTE = Int32(1)
@@ -127,26 +127,8 @@ const NPY_UNICODE = Int32(19)
 const NPY_VOID = Int32(20)
 const NPY_HALF = Int32(23)
 
-# NPY_ORDER:
-const NPY_ANYORDER = Int32(-1)
-const NPY_CORDER = Int32(0)
-const NPY_FORTRANORDER = Int32(1)
-
-# flags:
-const NPY_ARRAY_C_CONTIGUOUS = Int32(1)
-const NPY_ARRAY_F_CONTIGUOUS = Int32(2)
-const NPY_ARRAY_ALIGNED = Int32(0x0100)
-const NPY_ARRAY_WRITEABLE = Int32(0x0400)
-const NPY_ARRAY_OWNDATA = Int32(0x0004)
-const NPY_ARRAY_ENSURECOPY = Int32(0x0020)
-const NPY_ARRAY_ENSUREARRAY = Int32(0x0040)
-const NPY_ARRAY_FORCECAST = Int32(0x0010)
-const NPY_ARRAY_UPDATEIFCOPY = Int32(0x1000)
-const NPY_ARRAY_NOTSWAPPED = Int32(0x0200)
-const NPY_ARRAY_ELEMENTSTRIDES = Int32(0x0080)
-
 #########################################################################
-# conversion from Julia types to NPY_TYPES constant
+# conversion from Julia types to NumPy types
 
 npy_type(::Type{Bool}) = NPY_BOOL
 npy_type(::Type{Int8}) = NPY_BYTE
@@ -164,18 +146,9 @@ npy_type(::Type{ComplexF32}) = NPY_CFLOAT
 npy_type(::Type{ComplexF64}) = NPY_CDOUBLE
 npy_type(::Type{PyPtr}) = NPY_OBJECT
 
-const NPY_TYPES = Union{Bool,Int8,UInt8,Int16,UInt16,Int32,UInt32,Int64,UInt64,Float16,Float32,Float64,ComplexF32,ComplexF64,PyPtr}
-
-# conversions from __array_interface__ type strings to supported Julia types
-const npy_typestrs = Dict( "b1"=>Bool,
-                           "i1"=>Int8,        "u1"=>UInt8,
-                           "i2"=>Int16,       "u2"=>UInt16,
-                           "i4"=>Int32,       "u4"=>UInt32,
-                           "i8"=>Int64,       "u8"=>UInt64,
-                           "f2"=>Float16,     "f4"=>Float32,
-                           "f8"=>Float64,     "c8"=>ComplexF32,
-                           "c16"=>ComplexF64, "O"=>PyPtr,
-                           "O$(div(Sys.WORD_SIZE,8))"=>PyPtr)
+# flags:
+const NPY_ARRAY_ALIGNED = Int32(0x0100)
+const NPY_ARRAY_WRITEABLE = Int32(0x0400)
 
 #########################################################################
 # no-copy conversion of Julia arrays to NumPy arrays.
@@ -185,7 +158,7 @@ const npy_typestrs = Dict( "b1"=>Bool,
 # dimensions. For example, although NumPy works with both row-major and
 # column-major data, some Python libraries like OpenCV seem to require
 # row-major data (the default in NumPy). In such cases, use PyReverseDims(array)
-function NpyArray(a::StridedArray{T}, revdims::Bool) where T<:NPY_TYPES
+function NpyArray(a::StridedArray{T}, revdims::Bool) where T<:PYARR_TYPES
     @npyinitialize
     size_a = revdims ? reverse(size(a)) : size(a)
     strides_a = revdims ? reverse(strides(a)) : strides(a)
@@ -199,7 +172,7 @@ function NpyArray(a::StridedArray{T}, revdims::Bool) where T<:NPY_TYPES
     return PyObject(p, a)
 end
 
-function PyObject(a::StridedArray{T}) where T<:NPY_TYPES
+function PyObject(a::StridedArray{T}) where T<:PYARR_TYPES
     try
         return NpyArray(a, false)
     catch
@@ -207,7 +180,7 @@ function PyObject(a::StridedArray{T}) where T<:NPY_TYPES
     end
 end
 
-PyReverseDims(a::StridedArray{T}) where {T<:NPY_TYPES} = NpyArray(a, true)
+PyReverseDims(a::StridedArray{T}) where {T<:PYARR_TYPES} = NpyArray(a, true)
 PyReverseDims(a::BitArray) = PyReverseDims(Array(a))
 
 """
@@ -220,301 +193,5 @@ a 4×3×2 NumPy row-major array).  This is useful for Python
 libraries that expect row-major data.
 """
 PyReverseDims(a::AbstractArray)
-
-#########################################################################
-# Extract shape and other information about a NumPy array.  We need
-# to call the Python interface to do this, since the equivalent information
-# in NumPy's C API is only available via macros (or parsing structs).
-# [ Hopefully, this will be improved in a future NumPy version. ]
-struct PyArray_Info{T,N}
-    native::Bool # native byte order?
-    sz::NTuple{N,Int}
-    st::NTuple{N,Int} # strides, in multiples of bytes!
-    data::Ptr{T}
-    readonly::Bool
-end
-
-function PyArray_Info(a::PyObject)
-    ai = PyDict{String,PyObject,true}(a["__array_interface__"])
-    typestr = convert(String, ai["typestr"])
-    T = npy_typestrs[typestr[2:end]]
-    datatuple = convert(Tuple{Int,Bool}, ai["data"])
-    sz = convert(Tuple{Vararg{Int}}, ai["shape"])
-    N = length(sz)
-    st = if isempty(sz)
-        ()
-    else
-        stq = get(ai, "strides", PyNULL())
-        if stq.o == pynothing[] || stq.o == C_NULL
-            default_stride(sz, T)
-        else
-            convert(Tuple{Vararg{Int}}, stq)
-        end
-    end
-    return PyArray_Info{T,N}((ENDIAN_BOM == 0x04030201 && typestr[1] == '<')
-               || (ENDIAN_BOM == 0x01020304 && typestr[1] == '>')
-               || typestr[1] == '|',
-               sz, st,
-               convert(Ptr{Cvoid}, datatuple[1]),
-               datatuple[2])
-end
-
-aligned(i::PyArray_Info{T,N}) where {T,N} = #  FIXME: also check pointer alignment?
-  all(m -> m == 0, mod.(i.st, sizeof(T))) # strides divisible by elsize
-
-eltype(i::PyArray_Info{T,N}) where {T,N} = T
-ndims(i::PyArray_Info{T,N})  where {T,N} = N
-
-function default_stride(sz::NTuple{N, Int}, ::Type{T}) where {T,N}
-    stv = Vector{Int}(N)
-    stv[end] = sizeof(T)
-    for i = N-1:-1:1
-        stv[i] = stv[i+1]*sz[i+1]
-    end
-    ntuple(i->stv[i], N)
-end
-
-# whether a contiguous array in column-major (Fortran, Julia) order
-function f_contiguous(T::Type, sz::NTuple{N,Int}, st::NTuple{N,Int}) where N
-    if prod(sz) == 1 || length(sz) == 1
-        # 0 or 1-dim arrays should default to f-contiguous in julia
-        return true
-    end
-    if st[1] != sizeof(T)
-        return false
-    end
-    for j = 2:N
-        if st[j] != st[j-1] * sz[j-1]
-            return false
-        end
-    end
-    return true
-end
-
-f_contiguous(T::Type, sz::NTuple{N1,Int}, st::NTuple{N2,Int}) where {N1,N2} =
-    error("stride and size are different lengths, size: $sz, strides: $sz")
-
-f_contiguous(i::PyArray_Info{T,N}) where {T,N} = f_contiguous(T, i.sz, i.st)
-c_contiguous(i::PyArray_Info{T,N}) where {T,N} = f_contiguous(T, reverse(i.sz), reverse(i.st))
-
-#########################################################################
-# PyArray: no-copy wrapper around NumPy ndarray
-#
-# Hopefully, in the future this can be a subclass of StridedArray (see
-# Julia issue #2345), which will allow it to be used with most Julia
-# functions, but that is not possible at the moment.  So, to use this
-# with Julia linalg functions etcetera a copy is still required.
-
-"""
-    PyArray(o::PyObject)
-
-This converts an `ndarray` object `o` to a PyArray.
-
-This implements a nocopy wrapper to a NumPy array (currently of only numeric types only).
-
-If you are using `pycall` and the function returns an `ndarray`, you can use `PyArray` as the return type to directly receive a `PyArray`.
-"""
-mutable struct PyArray{T,N} <: AbstractArray{T,N}
-    o::PyObject
-    info::PyArray_Info
-    dims::Dims
-    st::NTuple{N,Int}
-    f_contig::Bool
-    c_contig::Bool
-    data::Ptr{T}
-
-    function PyArray{T,N}(o::PyObject, info::PyArray_Info) where {T,N}
-        if !aligned(info)
-            throw(ArgumentError("only NPY_ARRAY_ALIGNED arrays are supported"))
-        elseif !info.native
-            throw(ArgumentError("only native byte-order arrays are supported"))
-        elseif eltype(info) != T
-            throw(ArgumentError("inconsistent type in PyArray constructor"))
-        elseif length(info.sz) != N || length(info.st) != N
-            throw(ArgumentError("inconsistent ndims in PyArray constructor"))
-        end
-        return new{T,N}(o, info, tuple(info.sz...), div.(info.st, sizeof(T)),
-                        f_contiguous(info), c_contiguous(info),
-                        convert(Ptr{T}, info.data))
-    end
-end
-
-function PyArray(o::PyObject)
-    info = PyArray_Info(o)
-    return PyArray{eltype(info), length(info.sz)}(o, info)
-end
-
-size(a::PyArray) = a.dims
-ndims(a::PyArray{T,N}) where {T,N} = N
-
-similar(a::PyArray, T, dims::Dims) = Array{T}(undef, dims)
-
-function setdata!{T,N}(a::PyArray{T,N}, o::PyObject, pybufinfo=PyBuffer())
-    PyBuffer!(pybufinfo, o, PyBUF_ND_CONTIGUOUS)
-    dataptr = pybufinfo.buf.buf
-    a.data = reinterpret(Ptr{T}, dataptr)
-    a
-end
-
-function copy(a::PyArray{T,N}) where {T,N}
-    if N > 1 && a.c_contig # equivalent to f_contig with reversed dims
-        B = unsafe_wrap(Array, a.data, ntuple((n -> a.dims[N - n + 1]), N))
-        return permutedims(B, (N:-1:1))
-    end
-    A = Array{T}(undef, a.dims)
-    if a.f_contig
-        ccall(:memcpy, Cvoid, (Ptr{T}, Ptr{T}, Int), A, a, sizeof(T)*length(a))
-        return A
-    else
-        return copyto!(A, a)
-    end
-end
-
-# TODO: need to do bounds-checking of these indices!
-
-getindex(a::PyArray{T,0}) where {T} = unsafe_load(a.data)
-getindex(a::PyArray{T,1}, i::Integer) where {T} = unsafe_load(a.data, 1 + (i-1)*a.st[1])
-
-getindex(a::PyArray{T,2}, i::Integer, j::Integer) where {T} =
-  unsafe_load(a.data, 1 + (i-1)*a.st[1] + (j-1)*a.st[2])
-
-function getindex(a::PyArray, i::Integer)
-    if a.f_contig
-        return unsafe_load(a.data, i)
-    else
-        return a[ind2sub(a.dims, i)...]
-    end
-end
-
-function getindex(a::PyArray, is::Integer...)
-    index = 1
-    n = min(length(is),length(a.st))
-    for i = 1:n
-        index += (is[i]-1)*a.st[i]
-    end
-    for i = n+1:length(is)
-        if is[i] != 1
-            throw(BoundsError())
-        end
-    end
-    unsafe_load(a.data, index)
-end
-
-function writeok_assign(a::PyArray, v, i::Integer)
-    if a.info.readonly
-        throw(ArgumentError("read-only PyArray"))
-    else
-        unsafe_store!(a.data, v, i)
-    end
-    return a
-end
-
-setindex!(a::PyArray{T,0}, v) where {T} = writeok_assign(a, v, 1)
-setindex!(a::PyArray{T,1}, v, i::Integer) where {T} = writeok_assign(a, v, 1 + (i-1)*a.st[1])
-
-setindex!(a::PyArray{T,2}, v, i::Integer, j::Integer) where {T} =
-  writeok_assign(a, v, 1 + (i-1)*a.st[1] + (j-1)*a.st[2])
-
-function setindex!(a::PyArray, v, i::Integer)
-    if a.f_contig
-        return writeok_assign(a, v, i)
-    else
-        return setindex!(a, v, ind2sub(a.dims, i)...)
-    end
-end
-
-function setindex!(a::PyArray, v, is::Integer...)
-    index = 1
-    n = min(length(is),length(a.st))
-    for i = 1:n
-        index += (is[i]-1)*a.st[i]
-    end
-    for i = n+1:length(is)
-        if is[i] != 1
-            throw(BoundsError())
-        end
-    end
-    writeok_assign(a, v, index)
-end
-
-stride(a::PyArray, i::Integer) = a.st[i]
-
-Base.unsafe_convert(::Type{Ptr{T}}, a::PyArray{T}) where {T} = a.data
-
-pointer(a::PyArray, i::Int) = pointer(a, ind2sub(a.dims, i))
-
-function pointer(a::PyArray{T}, is::Tuple{Vararg{Int}}) where T
-    offset = 0
-    for i = 1:length(is)
-        offset += (is[i]-1)*a.st[i]
-    end
-    return a.data + offset*sizeof(T)
-end
-
-summary(a::PyArray{T}) where {T} = string(Base.dims2string(size(a)), " ",
-                                          string(T), " PyArray")
-
-#########################################################################
-# PyArray <-> PyObject conversions
-
-PyObject(a::PyArray) = a.o
-
-convert(::Type{PyArray}, o::PyObject) = PyArray(o)
-
-function convert(::Type{Array{T, 1}}, o::PyObject) where T<:NPY_TYPES
-    try
-        copy(PyArray{T, 1}(o, PyArray_Info(o))) # will check T and N vs. info
-    catch
-        len = @pycheckz ccall((@pysym :PySequence_Size), Int, (PyPtr,), o)
-        A = Array{pyany_toany(T)}(undef, len)
-        py2array(T, A, o, 1, 1)
-    end
-end
-
-function convert(::Type{Array{T}}, o::PyObject) where T<:NPY_TYPES
-    try
-        info = PyArray_Info(o)
-        try
-            copy(PyArray{T, length(info.sz)}(o, info)) # will check T == eltype(info)
-        catch
-            return py2array(T, Array{pyany_toany(T)}(undef, info.sz...), o, 1, 1)
-        end
-    catch
-        py2array(T, o)
-    end
-end
-
-function convert(::Type{Array{T,N}}, o::PyObject) where {T<:NPY_TYPES,N}
-    try
-        info = PyArray_Info(o)
-        try
-            copy(PyArray{T,N}(o, info)) # will check T,N == eltype(info),ndims(info)
-        catch
-            nd = length(info.sz)
-            if nd != N
-                throw(ArgumentError("cannot convert $(nd)d array to $(N)d"))
-            end
-            return py2array(T, Array{pyany_toany(T)}(undef, info.sz...), o, 1, 1)
-        end
-    catch
-        A = py2array(T, o)
-        if ndims(A) != N
-            throw(ArgumentError("cannot convert $(ndims(A))d array to $(N)d"))
-        end
-        A
-    end
-end
-
-function convert(::Type{Array{PyObject}}, o::PyObject)
-    map(pyincref, convert(Array{PyPtr}, o))
-end
-
-function convert(::Type{Array{PyObject,1}}, o::PyObject)
-    map(pyincref, convert(Array{PyPtr, 1}, o))
-end
-
-function convert(::Type{Array{PyObject,N}}, o::PyObject) where N
-    map(pyincref, convert(Array{PyPtr, N}, o))
-end
 
 #########################################################################