@stevengj 2018-04-04
That is, `PyArray` (the no-copy conversion type) should just use the buffer API.