@JobJob 2018-04-04
as in everything related to `__aray_interface__`