@stevengj 2018-04-04
Right.... for that we'd need to create an object that implements the buffer API.   One concern is that I don't know how well this would work with Python functions expecting a NumPy array.

Okay, it makes sense to replace the Python -> Julia array code with buffer code in this PR, but leave the Julia -> Python stuff using NumPy for now and work on that in a separate PR.