@JobJob 2018-04-04
The other issue with replacing `PyArray(o::PyObject)` with `PyArrayFromBuffer` is if something implements `__array_interface__` but not the buffer interface. Not sure which is more supported in the wild? But we could just add a fallback to `__array_interface__` if the buffer interface isn't implemented.