@stevengj 2018-04-04
> if something implements `__array_interface__` but not the buffer interface. 

I'm guessing that this never happens in practice.