@JobJob 2018-04-04
> I'm guessing that this never happens in practice.

Ok, let's not support it then unless someone tells us it does 😄