@JobJob 2018-04-08
What's the issue with having numpy as a test dependency? Seems that using numpy will allow us to cover more cases, more easily.

I thought I might be able to get by with python's [array module](https://docs.python.org/3/library/array.html), but then I did a little digging and found out that it doesn't support the buffer interface on Python 2.7 🤦. You can get typed byte data with [struct.pack](https://docs.python.org/2/library/struct.html#struct.pack), but you can't reshape a memoryview using `cast` in py2 (you can in Python 3). So, the only way I could see to create a multi-dim array without using numpy in py2.7 is messing around in ctypes.

It could be done, but would take effort, and since I'm guessing the majority of the arrays these functions will see in the wild will be coming from numpy anyway (?), could be good to test with it.