@stevengj 2018-04-08
NumPy as a test dependency is fine; we should definitely test with NumPy arrays and operations.