@JobJob 2018-08-07
Assuming the tests pass I think this is pretty good to go. (they pass for me on mac with py 2/3 julia 0.6.3/0.7-dev-xxx)

There are things in pep-3118 that it doesn't handle, but I think it handles all the cases of arrays (strides and types) that current master does.

Oh the one thing I wanted to ask about is there's a deprecation warning about `strides` on 0.7 - so pyarray should implement strides?