@stevengj 2018-08-17
Nightly failures on `master` should be fixed now if you want to rebase.