@JobJob 2018-09-13
I think this is good to go now. Unlikely this PR is causing the Conda failures on Windows.