@JobJob 2018-11-15
I think it could be related to 48d730f293974f67543f2d5ef76173e4752cc30f
I can replicate on 0.7 with master locally, but when I reset  to 17dcb37441df2e68661c197049d88437bd3e0b18 on the branch of this PR (array-perf on https://github.com/JobJob/PyCall.jl) things are fine - i.e. no segfault at the end of testing