@marius311 2019-01-23
Apologies, I haven't followed the discussion above, but I did bisect a failure of my code when upgrading to PyCall#master down to this commit. It seems that before this commit, 

```
julia> using PyPlot
julia> subplots(1, 1; squeeze=false)[2]
1×1 Array{PyCall.PyObject,2}:
 PyObject <matplotlib.axes._subplots.AxesSubplot object at 0x7f75f95c7610>
```

but after

```
julia> using PyPlot
julia> subplots(1, 1; squeeze=false)[2]
PyObject array([[<matplotlib.axes._subplots.AxesSubplot object at 0x7ff0846ed610>]],
      dtype=object)
```

Was this intended or not? 

In either case, is there a workaround to force conversion to a Julia `Matrix` in the second case?