@stevengj 2019-01-24
Probably not intended.  Have you tried `convert(Matrix, ans)` on the result?