@marius311 2019-01-24
Thanks. Neither of these two attempts on current master seem to work:

```
julia> convert(Matrix, subplots(1,1;squeeze=false)[2])
ERROR: MethodError: Cannot `convert` an object of type PyObject to an object of type Array{T,2} where T
Closest candidates are:
  convert(::Type{T<:Array}, ::AbstractArray) where T<:Array at array.jl:474
  convert(::Type{T<:AbstractArray}, ::T<:AbstractArray) where T<:AbstractArray at abstractarray.jl:14
  convert(::Type{T<:AbstractArray}, ::LinearAlgebra.Factorization) where T<:AbstractArray at /home/marius/src/julia/usr/share/julia/stdlib/v1.1/LinearAlgebra/src/factorization.jl:46
  ...
Stacktrace:
 [1] top-level scope at none:0

julia> PyArray(subplots(1,1;squeeze=false)[2])
ERROR: KeyError: key "O" not found
Stacktrace:
 [1] getindex at ./dict.jl:478 [inlined]
 [2] array_format(::PyBuffer) at /home/marius/.julia/dev/PyCall/src/pybuffer.jl:261
 [3] PyArray_Info(::PyObject) at /home/marius/.julia/dev/PyCall/src/pyarray.jl:17
 [4] PyArray(::PyObject) at /home/marius/.julia/dev/PyCall/src/pyarray.jl:123
 [5] top-level scope at none:0
```

But `permutedims(getindex.(Ref(_), 1:m, 1:n))` works fine so I'm happy with that workaround.